Fix test prediction intervals
